Fixed main menu, added level select, fixed UI, added Pause Menu and added exit to change levels
Fixed main menu, added level selection with saves, to save what level has player completed. Fixed UI so that it sticks to the sides and is more in place now. Added Pause Menu. Added exit to change levels. P.S. To remove all saves of completed levels press "M" on the start menu (when it shows play, settings and quit)

https://abusivekoala.atlassian.net/jira/software/projects/TR/boards/1?selectedIssue=TR-172

https://abusivekoala.atlassian.net/jira/software/projects/TR/boards/1?selectedIssue=TR-174

https://abusivekoala.atlassian.net/jira/software/projects/TR/boards/1?selectedIssue=TR-176